Make sure form is initiallised properly on reorder
If an existing item that hasn't been changed yet is moved, the form
is rendered. But in this case, the form initialisation doesnt seem to
work properly, probably due to the parent element being moved to a
different part of the tree.

I've refactored it so that the form rendering is taken care of in a
child element. If the form is ever re-mounted by React, the
initialisation code will be run again.

Fixes #20